requirements.txt
google-genai
aiogram>=3.0.0
pydantic-settings
yt-dlp